Move code verification out of page component for reset password
